requirements.txt
numpy==1.26.2
opencv-contrib-python==4.8.0.76
pandas==2.1.4
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
tzdata==2023.3
